azure-storage-mcp/pyproject.toml
[project]
name = "azure-storage-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "azure-identity>=1.21.0",
    "azure-mgmt-monitor>=6.0.2",
    "azure-mgmt-resource>=23.3.0",
    "azure-mgmt-storage>=22.2.0",
    "mcp[cli]>=1.6.0",
    "pandas>=2.2.3",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]
